requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
PyNaCl==1.3.0
dnspython==1.16.0
async-timeout==3.0.1
discord.py==0.16.12